print binary

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
 */
 void print_binary(unsigned long int n)
 {
-	unsigned int num;
+	unsigned long int num;
 
 	if ((n >> 1) >= 0)
 	{
